Move code to sessions page and delete newEngine component
